feat(atom/input): postal code + file renaming + folder structure
