+++ components/atom/input/src/index.scss
@@ -1,5 +1,4 @@
 @import '~@schibstedspain/theme-basic/lib/index';
 
 .sui-AtomInput {
-  // Do your magic
 }
