+++ components/atom/input/src/inputs/index.js
@@ -1,0 +1,7 @@
+import DateInput from './DateInput'
+import MaskedInput from './MaskedInput'
+
+export {
+  DateInput,
+  MaskedInput
+}
